make death tests threadsafe
this prepares for the 'use logger in INVARIANT, PRECONDITION macros'
since there we will need to shutdown the logger to ensure flushing
before terminating when a INVARIANT or PRECONDITION is violated.

If the death tests are not thread safe, they will hang if the
NES::Logger::shutdown() call is present in the macros.