fix(SequenceShredder): Third usage for tuple delimiter buffers.
Prior, we assumed that a buffer with a tuple delimiter had two uses (in the SequenceShredder), one to create the leading and one to create the trailing spanning tuple. However, there is the additional use of returning the buffer to just process the tuples between the first and last delimiter in the buffer. By assuming two uses, we could run into the case, where the thread that processes the buffer puts it into the SequenceShredder and before retrieving it again at the end, two other threads would create the leading/trailing spanning tuples, reduce the uses to 0 and move the buffer out of the SequenceShredder again. When the thread that the buffer belongs to then wants to access the buffer, it is no more available.
To fix this, we added three uses, two like before and additional one for processing the tuples inside of the buffer.