chore(SourceThread): NumSourceLocalBuffers via WorkerConfig.
The 'numberOfBuffersInSourceLocalBufferPool' option already existed in our WorkerConfig, but we never fed it into the SourceThread. With this commit, we pass it to the 'SourceProvider::lower'. The SourceProvider checks if the user specified a value for this specific source. If the user did, the source-specific configuration takes priority. Otherwise, we use the value from the WorkerConfig (which is NodeEngine-wide).