fix(SequenceShredder): fixed stagedBuffers access
Prior, it could happen that a thread A with a buffer that contains a tuple delimiter added its buffer to the stagedBuffers and set the respective bit. Then, two other threads would find the leading/trailing spanning tuples starting/ending in the buffer of thread A. Thus the bitmap of buffer A could be completed and the SequenceShredder could move the tail past the bitmap of the buffer of A, allowing other threads to overwrite the staged buffer of A. A would then either find no buffer, its own buffer, or the buffer of another thread in the staged buffers. Thread A could also determine wrong spanning tuples.

This commit adds a check to determine whether the sequence number of a thread is still valid, when it wants to access the staged buffers. If not, it checks if the SequenceShredder still owns its buffers and releases the ownership. Otherwise, it just returns its own staged buffer.

Bad sequence:
- Thread A: gets buffer with sequence number X
- Thread A: calls processSequenceNumber(X) of sequence shredder
- Thread A: sets the respective bit and, e.g., 3 uses
- Thread B: processes SN X-1, finds the ST (forwards) to X, reduces uses to 2
- Thread C: processes SN X+1, finds the ST (backwards) to X, reduces uses to 1
- Thread C: sets seenAndUsed bit for SN X (started ST with it)
- Thread D: determines that bitmap that X belongs to is complete
- Thread D: increases tail, resets bitmap
- Thread E: calls processSequenceNumber(Y) of sequence shredder
- Thread E: determines stagedBuffer index for Y that matches staged buffer index of X
    - for example: (3490 % 2048) = 1442 = (5538 % 2048)
- Thread E: processes SN Y, finds some ST, reduces uses to 0
- Thread A: processes SN X, finds no ST, reduces uses to -1, gets empty buffer (or buffer of other thread)