feat(FieldAccessFunction): abstract function for field access
Enables, e.g., binary formats that simply calculate the tuple/field offsets to use(and implement) such field access functions without runtime overhead. Prior, InputFormatter implementations needed to fill TupleBuffers with offsets to fields. The more general FieldAccessFunction simply demands that an InputFormatter returns an object that can return a pointer to a field in the tuple that the object currently points to and to increment the pointer to the beginning of the next tuple. This allows to 'FieldAccessFunctions' that are, e.g., purely algebraic, which is a useful, e.g., for our native format.