fix(Memory): Fixes race condition in FixedSizeBufferPool destroy
A scenario where buffers are recycled to the FixedSizeBufferPool while it
is destroyed could potentially leak buffers.

This commit introduces a lock which is held during destruction of the
FixedSizeBufferPool and prevents such concurrent access. All recycle
calls are either happening before the destruction, which causes them
to be released during the destruction. Or happen after the destruction
which allows them to be recycled directly to the parent buffer.