feat(Systest): Prints the average throughput in bytes and tuples per second
We calculate the average throughput per second by dividing the file size of all input sources by the total query runtime (incl. submission and unregistering).
For the no. processed tuples, we count the no. lines in the input file. As the data is written in the csv format, this is fine.