chore(PluginRegistrationUtil): simplified add_plugin function
The add_plugin function now only takes the name of the plugin and the name of the registry that the plugin should be registered in.
We add the source code of the plugins, in particular the file that contains the register function, via our 'add_source_files' call, like all other source files.