feat(SysTest): Configurable systest sources
Allows to configure sources and input formatters in systests by using the 'Attach' keyword, to attach a physical source to a logical source. The 'Attach' keyword allows to configure the source, the input formatter and whether data should be read from a file or inline from the test file.
Since we usually have a one-to-one relation between our logical and physical sources, it makes sense to provide a shorthand that creates a physical FileSource with CSV format and that takes an INLINE/FILE parameter and a subsequent sequence of inline tuples or a file path. The shorthand syntax very closely resembles our old syntax.