feat(SmallFilesTest): Format agnostic result checking
Prior, in the SmallFilesTest UnitTests we relied on the `toString()` method of the TestTupleBuffer to convert the result buffers to CSV, which we then could write to a file and compare with the input file.
This only work for CSV.
This commit introduces expected buffers in our binary format that we can compare the formatted data against, which works for all formats. Additionally, it introduces functionality to produce such expected buffers.

# Conflicts:
#	nes-input-formatters/tests/UnitTests/SmallFilesTest.cpp