feat(FieldOffsets): Supporting fields that require more two offsets
If fields are separated by exactly the size of a field delimiter, knowing the field delimiters (start and the end of the tuple) suffices to determine the beginning and ends of all fields.
If that is not the case, it is necessary to store two offsets per field, one to mark the start of the field and one to mark the end. This is sufficient for all fields that are contiguously in memory.