fix(Execution): Fixes race condition in Emit Operator
There was a race between signaling that a chunk was seen and accessing
the current chunk counter. This would cause duplicate sequence number
and chunk number pairs to be emitted.

This commit changes the order to first query the next chunk number
before signaling that a chunk number was seen.