refactor(Projection): Combine Projection and Map Logical Operator
Previously the Projection and Map logical operator would do two distinct
things. Projection was used to limit the available fields and a map
allows adding new fields into the schema.

The refactored version of the `ProjectionLogicalOperator` combines these
functionalities. Fundamentally a projection is now a `FieldIdentifier` and
a `LogicalFunction`. After applying the projection the output schema will
contain the new FieldIdentifiers based on `LogicalFunctions`. In the special
case of an Asterisk `*` the input schema is added to the output schema.

On the physical level the Projection is lowered into a scan (which projects all
relevant fields) followed by N `MapPhysicalOperators`.

chore(Tests): Adapts system tests to new projection

The refactor of the projection introduced three major changes:
1. Fields created by a projection are no longer prefixed with
   a source name.
2. Group Keys, Start, Stop are no longer implicitly projected.
3. Projections can not include fields created by the same projection.
   If a projection maps a field with the same name the old version will
   be used by projections in the same SELECT statement.