feat(Join Testing): Added additional join tests for small and large tests files
All small join tests are executed with both join variants.
Also, we have added a new group NestedLoopJoin. This group represents tests that contain multiple 100s of MBs of input data but are small enough so that even the slow NLJ finishes the test in minutes. This is especially important for the Nexmark_more_data, as I could not get the NLJ to finish in 5 Minutes for one query.