use cmake: local cpp stdlib and no debug for host triplet
The host triplet is used to build dependencies that are run on the host
during the build process (e.g. the protoc compiler).
Thus we already do not e.g. build them with sanitizers (since we don't
really care if protoc e.g. leaks).

With this change we also always compile host dependencies with the
"default" local cpp stdlib, to simplify adding a custom libcxx.

We also only build release versions of the host dependencies,
which reduces built time and saves space.