cmake: chose cpp stdlib in simpler, stricter way
Previously, USE_LIBCXX_IF_AVAILABLE tried to use libc++ but implicitely
switched to libstdc++ if compiling with libc++ didn't work, which in
itself was potentially confusing.

With this change, USE_CPP_STDLIB can be set to either "local" or
"libcxx". There is no implicit change if something doesn't work but
explicit error messages.