fix(Systest): throw exception on INVALID token
Problem: Given a test file with two queries. One passes both the parser and the test (query passes, results are correct). The other fails during parsing (invalid token). We never register the second query. The one query we registered passes, so the test succeeds.

Now, the 'getTokenIfValid()' function of the SystestParser would returns an INVALID token (instead of a nullopt) when it detects an INVALID token. This triggers an exception and the test fails.
Added an e2e test to confirm behavior.