feat(LogicalFunctions): Extensible logical functions
Prior, the AntrlSqlQueryPlanCreator had a hard-coded if case to check if a function identifier matches 'CONCAT'.
Now, we use the LogicalFunctionRegistry to check if it contains the function identifier and if it does, we create the logical function. As a result, sql query parsing does not depend on specific function names anymore.