fix(engine): Addresses flakey test
This commit fixes a flakey test, caused by a race between
the engine shutdown and the query status listener reporting
the running event. Although unlikely, it is possible for the
running query status report to be skipped during query shutdown
even if the query has already produced multiple tuplebuffer.

The fix waits for the running event to be reported.