fix(Execution): Circumvent race condition for setting cleanup function
Prior commits make it possible for a single aggregation build operator
and operator handler to be shared across multiple pipelines.

This causes a race condition in the AggregationBuild setup, which is called multiple
times and attempts to initialize the cleanup function concurrently.

This commit fixes the issue by moving the construction of the cleanup function
into the constructor of the AggregationBuild which is not called concurrently.