fix(Systest): Use CMake ExternalData dir
Previously the systest tool was relying on the fact
that CMake ExternalData would store its data in the CMake
build directory. However this is not the case if none of
the ExternalData references are actual references, but just
plain files. In that ExternalData will just use the orignal
path.

This commit changes the default `SYSTEST_EXTERNAL_DATA_DIR`
to point to what ever directy ExternalData uses.