fix(LockFreeSequenceShredder): Prototype passes DistanceToDelimiterTest.cpp
Problem fixed:
Thread A goes into 'setCompletedFlags' for SN X and copies/loads the last entry (completedTrailing is false).
Thread B, for the same SN X tries to claim it for a spanning tuple.
Thread B claims the SN and sets 'completedTrailing' to true.
Thread A sets 'completedLeading' to true.
Thread A assigns to the cell containing the value for SN X and overwrites 'completedTrailing'.