feat(Sources): Implements ingestion control
Previously, ingestion control was handled via a limited number of
buffers in a FixedSizeBufferPool. With the removal of the
FixedSizeBufferPool the ingestion control is shifted into the
query engine, and is controlled via a semaphore that has to
be acquired before adding a new task and is released upon task
completion.

Additionally, this introduces a cancellation point into the
sources emit function.