feat(SequenceShredder): Replaces with lock-free implementation
Prior, the sequence shredder used a central lock to coordinate between InputFormatterTasks.
Additionally, the sequence shredder used two bitmaps to track and assign responsibility to threads for processing spanning tuples.
Determining spanning tuples was the result of several bitwise-operations on snapshots of bitmap(s)

This commit introduces an entirely different implementation of the SequenceShredder.
The new implementation operates without a (central) lock.
Instead, it uses 64-bit atomics to synchronize between threads.
The advantage is that threads that are interested in different parts of the ring buffer (that tracks input formatter buffers) don't need to synchronize anymore.
Additionally, except for the 64-bit atomic, the new implementation does not use bitmaps anymore. Instead, it uses a vector, where each cell of the vector represents a sequence number. As a result, determining spanning tuples does not involve bitwise operations anymore (except for interacting with the 64-bit atomic).