chore(SequenceShredder): Creating valid dummy buffer
Prior, we created a 'null' TupleBuffer for the very first entry that triggers the leading spanning tuple of the very first tuple. Since it was a null buffer, we needed to check at several places, whether the buffer was a null buffer or not.
Now we use a temporary buffer manager to create a valid dummy buffer, making the checks unnecessary.