refactor(engine): add task completion callback
There were issues where the reference reducer was not
called as it was part of the (now called) success callback
which was only called if the task returned true.

This commit adds onComplete callback which is called
regardless. It uses absl::AnyInvocable for its enforced
move semantics which make it impossible for a callback
to be executed multiple times.

# The commit message #2 will be skipped:

# fixup! refactor(engine): add task completion callback