fix(SystestResultCheck): Allow comparison of schemas with duplicate fields
Fixes a bug that would incorrectly throw a Schema Mismatch error for systests using sinks with duplicate fields.
This was caused by the compareSchemas method only using the first of the duplicate fields of the actual schema to match the expected schema.
Thus, the other duplicates remained unmatched, causing the error.