feat(logging): Adds context to logger
This commit removes the `async_logger` which means all logging
happens synchronously! While this will impact performance, all
performance-critical builds have very limited logging.

This allows us to use spdlog mdc and thread locals to attach
meta-information to the logging output. I.e., we can attach the
actual ThreadName as well as the WorkerId to the logging message.

The `LogContext` is a RAII wrapper for adding additional context to
all log messages that happen within the thread-local lifetime of the
`LogContext`. For example, the query engine can attach a query and
pipeline id to each log message generated during the execution of a
task.

Signed-off-by: lukas schwerdtfeger <lukas.schwerdtfeger@gmail.com>