feat(grpc): enable health service
This commit enables the grpc health service, which
is provided by grpc out of the box. The health service
exposes a grpc endpoint that can be queried to detect
if the worker is running.

This commit installs the `grpc-health-probe` into the
single-node-worker and nebuli docker images. The probe
can be used to specify a running condition in
docker-compose or kubernetes.

Signed-off-by: lukas schwerdtfeger <lukas.schwerdtfeger@gmail.com>