fix(SystestResultCheck): Removed expectedToActualFieldMap in SystestResultCheck
This commit removes the expectedToActualFieldMap struct in SystestResultCheck, which was used to compare query results if the actual schema had a different field order then the expected schema.
Since this mapping approach fails when dealing with schemas with duplicated fields, we decided to remove the mapping.
If a schema mismatch is detected, we populate the error stream with the output of the newly added SchemaDiff helper and immediatly end the result check.
All functions that used the expectedToActualFieldMap struct now assume that the actual fields are in the right order.