fix(engine): potential deadlock waiting for pending tasks
Previously the query engine could run into a deadlock if
a pipeline is requested to be shutdown gracefully, and the
PendingPipelineStop task would find that the pending tasks
counter within the pipeline is not zero (a.k.a, there
exists a task that is using that pipeline). This can be
problematic if that task is within the admission queue and
is never even considered to be picked up by and of the
worker threads. In a single threaded environment this
clearly produces a deadlock, if the same thread repeatedly
attempts the same tasks, puts it into the internal queue
only to pick it up again in the next iteration. This
example can also happen in a multithreaded environment.

The previous solution was to recursively execute the next
task and keep the PendingPipelineStop on its stack.
However, this could lead to unpredictable stack depths.

The new solution uses the DelayedTaskSubmitter to delay
the graceful pipeline stop for a fixed period of time.

Signed-off-by: lukas schwerdtfeger <lukas.schwerdtfeger@gmail.com>