fix(ExecutionContext): centralizes arena offset calculation
Prior, we kept a separate size and pointer in the ArenaRef, which could get out of sync with the Arena. Now, we simply invoke the 'allocateMemory call of the Arena'. This means we always proxy call when allocating memory. If this turns out to be a performance problem, we can resort to inlining, or accessing the members of the Arena directly.