chore(AntlrSQLQueryPlanCreator): functions consume last N args from stack
Prior, (logical) functions took all arguments from the `functionBuilder`. This could cause a crash, if there were still functions from the outer scope of the function-to-build on the stack (Concat.test:03). Now, functions simply consume the last N functions from the stack, where N is determined by the size of their children, which should be exactly the size of the functions that they consume.