chore(Decoders): Adds functioning codec option to source configuration.
The codec string option provides a way to tell the system, that the source produces encoded data. The currently supported codecs can be found in the Codec enum in Decoder.hpp.
If a codec is given (so not None), the DecoderProvider provides a decoder implementation for the codec that the source thread routine uses to decode incoming data.