chore(Decoders): Uses emitAndProvide function in decode methods
With the emitAndProvide function that is passed as argument to the decode functions, the decode method itself is able to emit a decodedTupleBuffer. Furthermore, emitAndProvide optionally returns a new, empty TupleBuffer so that the decode method can continue decoding even if the initial decoded buffer was filled completely.
This way, we only need to call decode once per encoded buffer.