chore(SinkDescriptor): adds parser format to SinkDescriptor
A Sink should generally have an input format.
This change makes this explicit by adding it to the parameter map of the SinkDescriptor.
Additionally, it allows accessing the format, which in turn allows, e.g., the PipeliningPhase to check if a source pipeline is immediately followed by a sink pipeline with the same format, enabling the source the hand its input buffers to the sink.