chore(SequenceShredder): Allows lazily setting lastTupleOffset
The SequenceShredder now checks if the offset of the last tuple was set when searching for spanning tuples (STs).
That allows InputFormatIndexers that can't determine the offset of the last tuple during indexing, e.g., because they need to parse the buffer first, to lazily set the offset of the lastTupleOffset. Threads can only find STs that start in buffer X, if buffer X has the offset of the last tuple delimiter set.
Behavior essentially stays the same for InputFormatIndexers that can determine the offset during indexing.