fix(PipeliningPhase): reject tuples larger than buffer size
Prior, we would not check if the configured buffer sizes of pipelines were large enough to contain a single tuple.
If it wasn't it could lead to ugly errors, that would not even result in a crash in the worst case.
This commit adds a check that throws the new TuplesTooLargeForPipelineBufferSize exception if a pipeline was not configured with a large enough buffer size.
This fix still works if we introduce pipeline-specific buffer sizes. It should also be simple to access a physical memory layout to access the real physical 'tuple' size if we introduce that.