fix(PredictionCache): store keys and values as pointers to heap-allocated byte buffers
The last change replaced pointers to byte buffers as keys and values with vectors. Under that implementation we reinterpreted the reference to the PredictionCacheEntry to the vector pointer. However, that often led to an undefined behavior, as we could not guarantee that we will get a valid pointer to the vector's data. With this patch we fall back to using pointers to byte buffers as keys and values, but this time we allocate them on the heap by copying the contents of adapter->inputData and adapter->outputData. We change the implementation of PredictionCacheEntry in accordance with the rule of five