feat(Systest): Add file level sequential execution of queries
To run queries sequentially add the `SEQUENTIAL_EXECUTION` statement on a line. All queries parsed in the file until the next `SEQUENTIALL_EXECUTION` will be submitted to run sequentially by the runQueries method

To ensure this functionality within the benchmark mode, the order of queries to run is simply reversed. As the benchmark mode already runs single threaded and the reversed order is the order the queries are written in the file

Additionally, two new tests were added to the systest runner test. One checking for a regression to a bug found, where it was possible to submit a query with a non-existent dependency causing the systest runner to hang. And another to ensure that dependent queries are actually started / registered in the right order